requirements.txt
python-telegram-bot
huggingface_hub